+++ lib/nova/api/resource/bill.rb
@@ -2,16 +2,38 @@ module Nova
   module API
     module Resource
       class Bill < Nova::API::Base
+        module DOCUMENT_TYPE
+          INVOICE = 0
+          RECEIPT = 1
+          COUPON_TAX = 2
+          BILL = 3
+          BILLET = 4
+          CHEQUE = 5
+          INITIAL_BALANCE = 6
+          TRANSFERENCE = 7
+          CONTRACT = 8
+          OTHER = 9
+          STATEMENT = 10
+        end
+
+        module DUE_TYPE
+          MONTHLY = 0
+          BIWEEKLY = 1
+          TEN_DAYS = 2
+          FIXED = 3
+          WEEKLY = 4
+        end
+
         class Apportionment < Nova::API::Utils::BaseStruct
           ALLOWED_ATTRIBUTES = [:value, :apportionment_value_ids]
 
           attribute :apportionment_value_ids, Dry::Types['strict.array'].of(Dry::Types['coercible.integer'].optional)
-          attribute :value, Dry::Types['coercible.decimal']
+          attribute :value, Dry::Types['coercible.float']
         end
 
         ALLOWED_ATTRIBUTES = [
           :additional_information, :apportionments, :company_id, :date, :document_type, :document_number, :due_type, :financial_account_id,
-          :first_due_date, :forecast, :installments, :installments_number, :third_party_id, :total_value
+          :first_due_date, :forecast, :identifier, :installments, :installments_number, :third_party_id, :total_value
         ]
 
         attribute? :id, Dry::Types['coercible.integer'].optional
@@ -25,10 +47,11 @@ module Nova
         attribute :financial_account_id, Dry::Types['coercible.integer']
         attribute :first_due_date, Dry::Types['coercible.string'].constrained(format: DATE_REGEX)
         attribute :forecast, Dry::Types['strict.bool']
+        attribute? :identifier, Dry::Types['coercible.string'].optional
         attribute? :installments, Dry::Types['strict.array'].of(Nova::API::Resource::Installment).optional
         attribute :installments_number, Dry::Types['coercible.integer']
         attribute :third_party_id, Dry::Types['coercible.integer']
-        attribute :total_value, Dry::Types['coercible.decimal']
+        attribute :total_value, Dry::Types['coercible.float']
 
         def self.endpoint
           '/api/bills'