+++ lib/nova/api/resource/response/current_asset_statement.rb
@@ -3,7 +3,7 @@ module Nova
     module Resource
       module Response
         class CurrentAssetStatement < Nova::API::Utils::BaseStruct
-          attribute? :last_statement, Dry::Types['coercible.decimal'].optional
+          attribute? :last_statement, Dry::Types['coercible.float'].optional
           attribute? :write_offs, Dry::Types['strict.array'].of(Nova::API::Resource::WriteOff).optional
         end
       end