+++ lib/nova/api/search_params/bill.rb
@@ -1,0 +1,9 @@
+module Nova
+  module API
+    module SearchParams
+      class Bill < Nova::API::Utils::BaseStruct
+        attribute? :identifier, Dry::Types['coercible.string'].optional
+      end
+    end
+  end
+end