+++ spec/nova/api/resource/bank_spec.rb
@@ -13,7 +13,7 @@ RSpec.describe Nova::API::Resource::Bank do
     it { is_expected.to have_attribute(:company, Nova::API::Resource::Company.optional) }
     it { is_expected.to have_attribute(:active, Dry::Types['strict.bool'].optional) }
     it { is_expected.to have_attribute(:image, Dry::Types['coercible.string'].optional) }
-    it { is_expected.to have_attribute(:balance, Dry::Types['coercible.decimal'].optional) }
+    it { is_expected.to have_attribute(:balance, Dry::Types['coercible.float'].optional) }
 
     it { is_expected.to have_attribute(:agency, Dry::Types['coercible.string'].optional) }
     it { is_expected.to have_attribute(:current_account, Dry::Types['coercible.string'].optional) }