+++ spec/nova/api/resource/write_off_spec.rb
@@ -10,7 +10,7 @@ RSpec.describe Nova::API::Resource::WriteOff do
     subject { described_class }
 
     it { is_expected.to have_attribute(:date, Dry::Types['coercible.string'].constrained(format: described_class::DATE_REGEX).optional) }
-    it { is_expected.to have_attribute(:value, Dry::Types['coercible.decimal'].optional) }
+    it { is_expected.to have_attribute(:value, Dry::Types['coercible.float'].optional) }
     it { is_expected.to have_attribute(:financial_account, Nova::API::Resource::FinancialAccount.optional) }
     it { is_expected.to have_attribute(:third_party, Nova::API::Resource::ThirdParty.optional) }
   end