+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0] - 2021-03-12
+
+### Added
+
+- Bill document types constants
+
+- Bill due types constants
+
+- Identifier attribute to bill resource
+
+- Receivables search endpoint
+
+- Payables search endpoint
+
 ## [0.6.0] - 2021-03-10
 
 ### Added